requirements.txt
youtube-search-python==1.6.6
customtkinter==5.2.2
pyinstaller==6.10.0
pytubefix==10.3.6
requests==2.32.5
pyperclip==1.9.0
CTkToolTip==0.8
pillow==11.0.0
httpx==0.27.2